requirements.txt
agno
openai
python-dotenv
duckduckgo-search
pypdf
lancedb
streamlit
groq
pandas
tantivy
yfinance
fastapi[standard]
sqlalchemy
